requirements.txt
Flask==0.10.1
coverage==4.0.3
pep8==1.7.0
pylint==1.5.4
pytest==2.8.7
pytest-cov==2.2.1
requests==2.9.1
PyYAML==3.11
